pyproject.toml
[tool.poetry]
name = "application-reviewer"
version = "0.1.0"
description = ""
authors = ["Ali"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
httpx = "^0.23.3"
uvicorn = "^0.21.1"
certifi = "^2022.12.7"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
faker = "^18.3.1"
google-auth = "^2.16.3"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.82.0"
python-docx = "^0.8.11"
python-multipart = "^0.0.6"
aiofiles = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"

